+++ _includes/header.html
@@ -151,6 +151,24 @@
             </div>
         </header>
 
+        <section id="twitch-livestream-alert" class="container">
+            <div class="row">
+                <div class="col-12 col-md-10 offset-md-1">
+                    <div class="card last-updated text-center mt-3">
+                        <div class="card-body rounded">
+                            <p>
+                                <i id="live-icon" class="fa-solid fa-circle animated live infinite"></i>
+                                ¡Estoy <a href="/live/">en vivo</a>!
+                                <a href="/live/" data-toggle="tooltip" data-placement="top" title="Ver stream en vivo">
+                                    <img style="vertical-align: text-top;" alt="Twitch Status" src="https://img.shields.io/twitch/status/mijodido?style=flat&logo=twitch&logoColor=%23A172F7&link=https%3A%2F%2Fluiscarlospando.com%2Flive">
+                                </a>
+                            </p>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </section>
+
         <section id="m7gp-livestream-alert" class="container">
             <div class="row">
                 <div class="col-12 col-md-10 offset-md-1">
