+++ _sass/header.scss
@@ -171,6 +171,7 @@
     }
 }
 
+#twitch-livestream-alert,
 #m7gp-livestream-alert {
     opacity: 0;
 
@@ -235,6 +236,7 @@ progress::-moz-progress-bar {
     }
 }
 
+#twitch-livestream-alert,
 #m7gp-livestream,
 #btn-m7gp {
     display: none;
