+++ assets/src/app.js
@@ -15,6 +15,8 @@ import './last-updated-at';
 import './wp-api';
 import './lastfm-fetch';
 import './now';
+import './twitch-live';
+import './m7gp-live';
 import './main';
 
 /* React components */
