+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "bundle exec jekyll serve",
     "dev": "webpack serve",
-    "build": "webpack",
+    "build": "webpack --watch",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
